Improve debug system
If the string you search for is not found,
a seperate error is thrown to help debugging.

Additionally, the error output is cleaner to
read and understand.

In some cases, errors are expected, and it is important
that the user can comprehend what when wrong when
running the script from the shell.